Allow replacing default Casting implementation.
